Begun work on widget.
How to get a widget to write to a database without an active activity?
Do I need an Android service or can I pass pending intents to be written
next time the activity starts? The latter would require refactoring the
fastTrack method to accept a timestamp parameter instead of calling for
one at the time of execution.