First steps for filtering clicks.
We're creating an AlertDialog with a list of agents to check or uncheck.
We added a column to the Agents table to remember if we're filtering it.
Currently the agents are showing up in the alert dialog but checking
them doesn't seem to do anything. OH! I didn't add the Where clause to
the getAllClicks method. That's up next.