Agents now have rules!
Added a rules table to the database. The idea is that there is an
adversary, played by RuleJudge, who, whnever an agent clicks, sets up
alarms to reset the agent. Currently only time elapsed rule is
available, but it's set up for an absolute time rule to work too -
represented in millis from day start I think.