Multiple Widgets Bug Fix
PendingIntents need to be called uniquely to be different. If they only
differ in an extra, as ours used to, retrieving it twice retrieves the
same PendingIntent and updates the extras. We needed to have a unique
request code in our getService method, for which we used the agentId.
Thus we get unique PendingIntents and so multiple widgets can be clicked
and alarmed independently.